Moved hlc icon from logo folder into icon folder.
